Revert "Register Torchvision Ops as Cutom Ops (#1267)" (#1316)
This reverts commit 78f169b5e1b96871936e45ca310788f3b3e4bb78.

+++ torchvision/csrc/nms.h
@@ -8,7 +8,7 @@
 at::Tensor nms(
     const at::Tensor& dets,
     const at::Tensor& scores,
-    const double iou_threshold) {
+    const float iou_threshold) {
   if (dets.device().is_cuda()) {
 #ifdef WITH_CUDA
     if (dets.numel() == 0) {
